Move swept_volume out of copyleft to non-copyleft marching cubes
swept_volume only lived under copyleft because it used libigl's GPL
copyleft marching cubes for mesh extraction. Switch to libigl's
non-copyleft igl::marching_cubes (drop-in sparse overload), move the
sources and binding into the non-copyleft module, and expose
swept_volume in the main gpytoolbox namespace. The copyleft submodule
re-exports it for backwards compatibility.

Fixes #165
