+++ CMakeLists.txt
@@ -84,6 +84,11 @@ add_library(cpytoolbox
 	src/cpp/remesher/tangential_relaxation.h
 	src/cpp/microstl/microstl_wrappers.h
 	src/cpp/tinyply/tinyply_wrappers.h
+	src/cpp/swept_volume/fd_interpolate.h
+	src/cpp/swept_volume/gradient_descent_test.h
+	src/cpp/swept_volume/random_points_on_mesh.h
+	src/cpp/swept_volume/sparse_continuation.h
+	src/cpp/swept_volume/swept_volume.h
 	# Source
 	src/cpp/upper_envelope.cpp
 	src/cpp/ray_mesh_intersect_aabb.cpp
@@ -104,27 +109,15 @@ add_library(cpytoolbox
 	src/cpp/remesher/tangential_relaxation.cpp
 	src/cpp/microstl/microstl_wrappers.cpp
 	src/cpp/tinyply/tinyply_wrappers.cpp
-	# microstl
-	src/cpp/microstl/microstl.h
-	# tinyply
-	src/cpp/tinyply/tinyply.h
-	)
-
-	add_library(cpytoolbox_copyleft
-	STATIC
-	# SHARED
-	# Headers
 	src/cpp/swept_volume/fd_interpolate.cpp
 	src/cpp/swept_volume/gradient_descent_test.cpp
 	src/cpp/swept_volume/random_points_on_mesh.cpp
 	src/cpp/swept_volume/sparse_continuation.cpp
 	src/cpp/swept_volume/swept_volume.cpp
-	# Source
-	src/cpp/swept_volume/fd_interpolate.h
-	src/cpp/swept_volume/gradient_descent_test.h
-	src/cpp/swept_volume/random_points_on_mesh.h
-	src/cpp/swept_volume/sparse_continuation.h
-	src/cpp/swept_volume/swept_volume.h
+	# microstl
+	src/cpp/microstl/microstl.h
+	# tinyply
+	src/cpp/tinyply/tinyply.h
 	)
 
 target_link_libraries(cpytoolbox ${LIBRARIES_TO_LINK})
@@ -135,8 +128,6 @@ if (CMAKE_CXX_COMPILER_ID STREQUAL "Clang")
   endif()
 endif()
 set(LIBRARIES_TO_LINK cpytoolbox ${LIBRARIES_TO_LINK})
-target_link_libraries(cpytoolbox_copyleft ${COPYLEFT_LIBRARIES_TO_LINK})
-set(COPYLEFT_LIBRARIES_TO_LINK cpytoolbox_copyleft ${COPYLEFT_LIBRARIES_TO_LINK})
 
 # Otman's python bindings using pybind11
 add_subdirectory(./ext/pybind11/)
@@ -149,6 +140,7 @@ pybind11_add_module(gpytoolbox_bindings
 	"${CMAKE_CURRENT_SOURCE_DIR}/src/cpp/binding_hausdorff_distance.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/src/cpp/binding_in_element_aabb.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/src/cpp/binding_marching_cubes.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/src/cpp/binding_swept_volume.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/src/cpp/binding_offset_surface.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/src/cpp/binding_point_mesh_squared_distance.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/src/cpp/binding_ray_mesh_intersect.cpp"
@@ -167,7 +159,6 @@ pybind11_add_module(gpytoolbox_bindings
 
 pybind11_add_module(gpytoolbox_bindings_copyleft
     "${CMAKE_CURRENT_SOURCE_DIR}/src/cpp/gpytoolbox_bindings_copyleft_core.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/src/cpp/binding_swept_volume.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/src/cpp/binding_booleans.cpp"
 )
