+++ src/gpytoolbox/copyleft/__init__.py
@@ -1,4 +1,6 @@
 from .mesh_boolean import mesh_boolean
 from .lazy_cage import lazy_cage
 from .do_meshes_intersect import do_meshes_intersect
-from .swept_volume import swept_volume
+# swept_volume no longer relies on copyleft (GPL) code; it now lives in the main
+# gpytoolbox namespace. Re-exported here for backwards compatibility.
+from gpytoolbox.swept_volume import swept_volume
